requirements.txt
asgiref==3.8.1
asttokens 
certifi==2024.8.30
charset-normalizer==3.4.0
decorator 
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
executing 
idna==3.10
ipython 
jedi 
matplotlib-inline 
parso 
pexpect 
prompt-toolkit 
ptyprocess 
pure-eval 
Pygments 
requests==2.32.3
six 
sqlparse==0.5.2
stack-data 
traitlets 
typing_extensions 
urllib3==2.2.3
wcwidth 
